LPS-201420 - Move _getBranchNavItems from NavigationMenuTag to NavItemUtil.
